+++ rs/tests/networking/nns_delegation_test.rs
@@ -18,8 +18,13 @@ Success::
 . NNS subnet doesn't attach any delegations to the responses.
 . Application subnets refresh delegations once in a while.
 . Responses to `api/v2/subnet/{subnet_id}/read_state` have valid delegations with canister ranges in the flat format.
+. Responses to `api/v3/subnet/{subnet_id}/read_state` have valid delegations without any canister ranges.
 . Responses to `api/v2/canister/{canister_id}/read_state` have valid delegations with canister ranges in the flat format.
+. Responses to `api/v3/canister/{canister_id}/read_state` have valid delegations with canister ranges in the tree format.
 . Responses to `api/v3/canister/{canister_id}/call` have valid delegations with canister ranges in the flat format.
+. Responses to `api/v4/canister/{canister_id}/call` have valid delegations with canister ranges in the tree format.
+. For `api/v2/canister/{canister_id}/query` we pass valid delegations with canister ranges in the flat format to the canister.
+. For `api/v3/canister/{canister_id}/query` we pass valid delegations with canister ranges in the tree format to the canister.
  */
 use std::{
     borrow::Cow,
@@ -27,11 +32,12 @@ use std::{
 };
 
 use anyhow::Result;
+use candid::{CandidType, Encode};
 use ic_agent::{
     agent::{Envelope, EnvelopeContent},
     Agent, Identity,
 };
-use ic_certification::validate_subnet_delegation_certificate;
+use ic_certification::verify_delegation_certificate;
 use ic_consensus_system_test_utils::{
     rw_message::install_nns_and_check_progress,
     upgrade::{
@@ -56,9 +62,13 @@ use ic_system_test_driver::{
     util::{block_on, get_identity, get_nns_node, UniversalCanister},
 };
 use ic_types::{
-    messages::{Blob, Certificate, CertificateDelegation, HttpReadStateResponse},
-    Height, PrincipalId, SubnetId,
+    messages::{
+        Blob, Certificate, CertificateDelegation, CertificateDelegationFormat,
+        HttpQueryResponseReply, HttpReadStateResponse, NodeSignature,
+    },
+    CanisterId, Height, PrincipalId, SubnetId,
 };
+use ic_universal_canister::wasm;
 use reqwest::StatusCode;
 use serde::{Deserialize, Serialize};
 use slog::info;
@@ -219,7 +229,30 @@ fn subnet_read_state_v2_returns_correct_delegation(env: TestEnv) {
             .delegation
             .expect("Should have an NNS delegation attached"),
         subnet.subnet_id,
-        Some(CanisterRangesFormat::Flat),
+        None,
+        CertificateDelegationFormat::Flat,
+    );
+}
+
+/// Responses to `api/v3/subnet/{subnet_id}/read_state` have valid delegations with canister ranges in the flat format.
+fn subnet_read_state_v3_returns_correct_delegation(env: TestEnv) {
+    let (subnet, node) = get_subnet_and_node(&env, SubnetType::Application);
+
+    let response: HttpReadStateResponse = block_on(send(
+        &node,
+        format!("api/v3/subnet/{}/read_state", subnet.subnet_id),
+        sign_envelope(&read_state_content()),
+    ));
+    let certificate: Certificate = serde_cbor::from_slice(&response.certificate).unwrap();
+
+    validate_delegation(
+        &env,
+        &certificate
+            .delegation
+            .expect("Should have an NNS delegation attached"),
+        subnet.subnet_id,
+        None,
+        CertificateDelegationFormat::Pruned,
     );
 }
 
@@ -243,7 +276,54 @@ fn canister_read_state_v2_returns_correct_delegation(env: TestEnv) {
             .delegation
             .expect("Should have an NNS delegation attached"),
         subnet.subnet_id,
-        Some(CanisterRangesFormat::Flat),
+        Some(node.effective_canister_id()),
+        CertificateDelegationFormat::Flat,
+    );
+}
+
+/// Responses to `api/v3/canister/{canister_id}/read_state` have valid delegations with canister ranges in the flat format.
+fn canister_read_state_v3_returns_correct_delegation(env: TestEnv) {
+    let (subnet, node) = get_subnet_and_node(&env, SubnetType::Application);
+
+    let response: HttpReadStateResponse = block_on(send(
+        &node,
+        format!(
+            "api/v3/canister/{}/read_state",
+            node.effective_canister_id()
+        ),
+        sign_envelope(&read_state_content()),
+    ));
+    let certificate: Certificate = serde_cbor::from_slice(&response.certificate).unwrap();
+
+    validate_delegation(
+        &env,
+        &certificate
+            .delegation
+            .expect("Should have an NNS delegation attached"),
+        subnet.subnet_id,
+        Some(node.effective_canister_id()),
+        CertificateDelegationFormat::Tree,
+    );
+}
+
+/// Responses to `api/v3/canister/aaaaa-aa/read_state` have valid delegations without canister ranges.
+fn canister_read_state_v3_management_canister_returns_correct_delegation(env: TestEnv) {
+    let (subnet, node) = get_subnet_and_node(&env, SubnetType::Application);
+    let response: HttpReadStateResponse = block_on(send(
+        &node,
+        format!("api/v3/canister/{}/read_state", CanisterId::ic_00()),
+        sign_envelope(&read_state_content()),
+    ));
+    let certificate: Certificate = serde_cbor::from_slice(&response.certificate).unwrap();
+
+    validate_delegation(
+        &env,
+        &certificate
+            .delegation
+            .expect("Should have an NNS delegation attached"),
+        subnet.subnet_id,
+        None,
+        CertificateDelegationFormat::Pruned,
     );
 }
 
@@ -271,8 +351,136 @@ fn call_v3_returns_correct_delegation(env: TestEnv) {
             .delegation
             .expect("Should have an NNS delegation attached"),
         subnet.subnet_id,
-        Some(CanisterRangesFormat::Flat),
+        Some(node.effective_canister_id()),
+        CertificateDelegationFormat::Flat,
     );
+}
+
+/// Responses to `api/v4/canister/{canister_id}/call` have valid delegations with canister ranges in the flat format.
+fn call_v4_returns_correct_delegation(env: TestEnv) {
+    let (subnet, node) = get_subnet_and_node(&env, SubnetType::Application);
+
+    let response: SyncCallResponse = block_on(send(
+        &node,
+        format!("api/v4/canister/{}/call", node.effective_canister_id()),
+        sign_envelope(&call_content(node.effective_canister_id())),
+    ));
+    let certificate: Certificate = serde_cbor::from_slice(&response.certificate).unwrap();
+
+    validate_delegation(
+        &env,
+        &certificate
+            .delegation
+            .expect("Should have an NNS delegation attached"),
+        subnet.subnet_id,
+        Some(node.effective_canister_id()),
+        CertificateDelegationFormat::Tree,
+    );
+}
+
+/// Responses to `api/v4/canister/{canister_id}/call` targeting the management canister
+/// have valid delegations without canister ranges.
+fn call_v4_management_canister_returns_correct_delegation(env: TestEnv) {
+    let (subnet, node) = get_subnet_and_node(&env, SubnetType::Application);
+    let expiration = OffsetDateTime::now_utc() + Duration::from_secs(3 * 60);
+
+    #[derive(CandidType)]
+    struct Arg {
+        canister_id: PrincipalId,
+    }
+    // This update call is a no-op because the canister is already started, but at least
+    // the call doesn't fail.
+    let call_content = EnvelopeContent::Call {
+        ingress_expiry: expiration.unix_timestamp_nanos() as u64,
+        sender: get_identity().sender().unwrap(),
+        canister_id: CanisterId::ic_00().into(),
+        method_name: String::from("start_canister"),
+        arg: Encode!(&Arg {
+            canister_id: node.effective_canister_id()
+        })
+        .unwrap(),
+        nonce: None,
+    };
+
+    let response: SyncCallResponse = block_on(send(
+        &node,
+        format!("api/v4/canister/{}/call", node.effective_canister_id()),
+        sign_envelope(&call_content),
+    ));
+    let certificate: Certificate = serde_cbor::from_slice(&response.certificate).unwrap();
+
+    validate_delegation(
+        &env,
+        &certificate
+            .delegation
+            .expect("Should have an NNS delegation attached"),
+        subnet.subnet_id,
+        Some(node.effective_canister_id()),
+        CertificateDelegationFormat::Tree,
+    );
+}
+
+#[derive(Deserialize)]
+struct QueryResponse {
+    reply: HttpQueryResponseReply,
+    #[allow(dead_code)]
+    signatures: Vec<NodeSignature>,
+}
+
+/// For `api/v2/canister/{canister_id}/query` we pass valid delegations with
+/// canister ranges in the flat format to the canister.
+fn query_v2_passes_correct_delegation_to_canister(env: TestEnv) {
+    let (subnet, node) = get_subnet_and_node(&env, SubnetType::Application);
+    let arg = wasm().data_certificate().append_and_reply().build();
+
+    let response: QueryResponse = block_on(send(
+        &node,
+        format!("api/v2/canister/{}/query", node.effective_canister_id()),
+        sign_envelope(&query_content(node.effective_canister_id(), arg)),
+    ));
+    let certificate: Certificate = serde_cbor::from_slice(&response.reply.arg).unwrap();
+
+    validate_delegation(
+        &env,
+        &certificate
+            .delegation
+            .expect("Should have an NNS delegation attached"),
+        subnet.subnet_id,
+        Some(node.effective_canister_id()),
+        CertificateDelegationFormat::Flat,
+    );
+}
+
+/// For `api/v3/canister/{canister_id}/query` we pass valid delegations with
+/// canister ranges in the tree format to the canister.
+fn query_v3_passes_correct_delegation_to_canister(env: TestEnv) {
+    let (subnet, node) = get_subnet_and_node(&env, SubnetType::Application);
+    let arg = wasm().data_certificate().append_and_reply().build();
+
+    let response: QueryResponse = block_on(send(
+        &node,
+        format!("api/v3/canister/{}/query", node.effective_canister_id()),
+        sign_envelope(&query_content(node.effective_canister_id(), arg)),
+    ));
+    let certificate: Certificate = serde_cbor::from_slice(&response.reply.arg).unwrap();
+
+    validate_delegation(
+        &env,
+        &certificate
+            .delegation
+            .expect("Should have an NNS delegation attached"),
+        subnet.subnet_id,
+        Some(node.effective_canister_id()),
+        CertificateDelegationFormat::Tree,
+    );
+}
+
+/// Run query tests several times sequentially to check that we don't return incorrect cached response.
+fn interlaced_v2_and_v3_query_requests(env: TestEnv) {
+    for _ in 0..10 {
+        query_v2_passes_correct_delegation_to_canister(env.clone());
+        query_v3_passes_correct_delegation_to_canister(env.clone());
+    }
 }
 
 fn read_state_content() -> EnvelopeContent {
@@ -296,6 +504,18 @@ fn call_content(canister_id: PrincipalId) -> EnvelopeContent {
     }
 }
 
+fn query_content(canister_id: PrincipalId, arg: Vec<u8>) -> EnvelopeContent {
+    let expiration = OffsetDateTime::now_utc() + Duration::from_secs(3 * 60);
+    EnvelopeContent::Query {
+        ingress_expiry: expiration.unix_timestamp_nanos() as u64,
+        sender: get_identity().sender().unwrap(),
+        canister_id: canister_id.into(),
+        method_name: String::from("query"),
+        arg,
+        nonce: None,
+    }
+}
+
 fn sign_envelope(content: &EnvelopeContent) -> Vec<u8> {
     let signature = get_identity().sign(content).unwrap();
 
@@ -314,23 +534,22 @@ fn sign_envelope(content: &EnvelopeContent) -> Vec<u8> {
     serialized_bytes
 }
 
-enum CanisterRangesFormat {
-    Flat,
-    #[allow(dead_code)]
-    Tree,
-}
-
 fn validate_delegation(
     env: &TestEnv,
     delegation: &CertificateDelegation,
     subnet_id: SubnetId,
-    canister_ranges_format: Option<CanisterRangesFormat>,
+    effective_canister_id: Option<PrincipalId>,
+    expected_delegation_format: CertificateDelegationFormat,
 ) {
     let nns_public_key = env.prep_dir("").unwrap().root_public_key().unwrap();
-    validate_subnet_delegation_certificate(
+    verify_delegation_certificate(
         &delegation.certificate,
         &subnet_id,
         &parse_threshold_sig_key_from_der(&nns_public_key).unwrap(),
+        effective_canister_id
+            .map(CanisterId::unchecked_from_principal)
+            .as_ref(),
+        /*use_signature_cache=*/ false,
     )
     .expect("Should receive a valid delegation certificate: {err:?}");
 
@@ -362,21 +581,25 @@ fn validate_delegation(
         });
 
     match (
-        canister_ranges_format,
+        expected_delegation_format,
         flat_canister_ranges,
         tree_canister_ranges,
     ) {
-        (None, None, None) => (),
-        (Some(CanisterRangesFormat::Tree), None, Some(_)) => (),
-        (Some(CanisterRangesFormat::Flat), Some(_), None) => (),
-        (None, Some(_), _) => panic!("Should not have any canister ranges"),
-        (None, _, Some(_)) => panic!("Should not have any canister ranges"),
-        (Some(CanisterRangesFormat::Flat), None, _) => panic!("Flat canister ranges not found"),
-        (Some(CanisterRangesFormat::Tree), _, None) => panic!("Tree canister ranges not found"),
-        (Some(CanisterRangesFormat::Tree), Some(_), _) => {
+        (CertificateDelegationFormat::Pruned, None, None) => (),
+        (CertificateDelegationFormat::Tree, None, Some(_)) => (),
+        (CertificateDelegationFormat::Flat, Some(_), None) => (),
+        (CertificateDelegationFormat::Pruned, Some(_), _) => {
+            panic!("Should not have any canister ranges")
+        }
+        (CertificateDelegationFormat::Pruned, _, Some(_)) => {
+            panic!("Should not have any canister ranges")
+        }
+        (CertificateDelegationFormat::Flat, None, _) => panic!("Flat canister ranges not found"),
+        (CertificateDelegationFormat::Tree, _, None) => panic!("Tree canister ranges not found"),
+        (CertificateDelegationFormat::Tree, Some(_), _) => {
             panic!("Should not have the flat canister ranges")
         }
-        (Some(CanisterRangesFormat::Flat), _, Some(_)) => {
+        (CertificateDelegationFormat::Flat, _, Some(_)) => {
             panic!("Should not have the tree canister ranges")
         }
     }
@@ -431,18 +654,16 @@ where
     let response = response
         .bytes()
         .await
-        .map_err(|err| format!("Request failed: {err}"))?
-        .to_vec();
+        .map_err(|err| format!("Request failed: {err}"))?;
 
     if status != StatusCode::OK {
         return Err(format!(
-            "Request failed. Status: {status}. Response: {}",
-            String::from_utf8_lossy(&response)
+            "Request failed. Status: {status}. Response: {response:?}",
         ));
     }
 
     serde_cbor::from_slice(&response)
-        .map_err(|err| format!("Failed to deserialize response: {err}"))
+        .map_err(|err| format!("Failed to deserialize response: {err:?}. Response: {response:?}",))
 }
 
 fn upgrade_application_subnet_if_necessary(env: &TestEnv) {
@@ -493,7 +714,20 @@ fn main() -> Result<()> {
         .add_test(systest!(nns_delegation_on_nns_test))
         .add_test(systest!(nns_delegation_on_app_subnet_test))
         .add_test(systest!(canister_read_state_v2_returns_correct_delegation))
+        .add_test(systest!(canister_read_state_v3_returns_correct_delegation))
+        .add_test(systest!(
+            canister_read_state_v3_management_canister_returns_correct_delegation
+        ))
         .add_test(systest!(subnet_read_state_v2_returns_correct_delegation))
+        .add_test(systest!(subnet_read_state_v3_returns_correct_delegation))
+        // note: the v2 call endpoint doesn't return an NNS delegation, so there is nothing to test
         .add_test(systest!(call_v3_returns_correct_delegation))
+        .add_test(systest!(call_v4_returns_correct_delegation))
+        .add_test(systest!(
+            call_v4_management_canister_returns_correct_delegation
+        ))
+        .add_test(systest!(query_v2_passes_correct_delegation_to_canister))
+        .add_test(systest!(query_v3_passes_correct_delegation_to_canister))
+        .add_test(systest!(interlaced_v2_and_v3_query_requests))
         .execute_from_args()
 }